pyproject.toml
[project]
name = "research-cnn"
version = "0.0.1"
description = "Research framework and utilities for studying Convolutional Neural Networks"
authors = [{name = "Alberto Perea León", email = "eralon.dev@gmail.com"}]
license = {text = "MIT"}
readme = "README.md"
requires-python = ">=3.10"

# Core runtime dependencies
dependencies = [
    "tensorflow>=2.17",     # ML Framework
    "numpy>=1.26",          # Tensor algebra helpers
    "pandas>=2.2.3",        # Data wrangling
    "polars>=1.3",          # Data wrangling
    "matplotlib>=3.9",      # Plotting
    "scikit-learn>=1.4",    # Metrics & helpers
    "jupyterlab>=4.1",      # Interactive notebooks
    "jupytext>=1.16",       # Notebook<->script pairing
    "shap>=0.47",           # Output ML explainer
    "pyyaml>=6.0.2",        # Config file parsing
    "tqdm>=4.67.1",           # Progress bars
]

[project.urls]
Homepage = "https://github.com/perealberto/cnn_research"

[project.optional-dependencies]
# Extras useful during development
dev = [
    "black>=25.1",          # Code formatter
    "nbstripout>=0.8",      # Clean notebook outputs before commit
    "pre-commit>=3.7",      # Git hooks manager
]

[build-system]
requires = ["setuptools >= 77.0.3"]
build-backend = "setuptools.build_meta"

[tool.hatch.metadata]
allow-direct-references = true

[tool.black]
line-length = 88
target-version = ["py311"]